feat(文件管理): 新增文件管理模块及相关功能
实现文件上传、下载、列表展示、标签管理等功能
添加文件管理路由和页面组件
更新配置文件支持文件管理接口
优化文件上传组件返回路径格式
新增文件大小格式化工具函数

+++ server/src/modules/user_admin/user_admin.service.ts
@@ -160,7 +160,7 @@ export class UserAdminService {
     } catch (e) {
       await fs.promises.mkdir(dir, { recursive: true });
     }
-    await fs.promises.writeFile(path.join(dir, Key), file.buffer);
+    await fs.promises.writeFile(path.join(dir, Key), file.buffer as Uint8Array);
     return `/uploadfile/${dirPath}/${Key}`;
   }
 }
